// generated/api/apiSchemas.ts
export type EntityID = string;
// ...
export type Jwk = {
  /**
   * The key type (e.g., EC).
   *
   * @example EC
   */
  kty: string;
  /**
   * The elliptic curve used (only for EC keys).
   *
   * @example P-256
   */
  crv?: string | null;
  /**
   * The key ID.
   *
   * @example 12345
   */
  kid: string;
  /**
   * The X coordinate for EC keys (optional).
   *
   * @example o-7zraXKDaoBte2PsuTXo-MSLzsyWdAElNptGgI4aH8
   */
  x?: string | null;
  /**
   * The Y coordinate for EC keys (optional).
   *
   * @example Xr_wCzJ1XnsgAIV5qHruzSwaNnwy87UjmevVklTpIv8
   */
  y?: string | null;
  /**
   * The modulus for RSA keys.
   *
   * @example modulus_value
   */
  n?: string | null;
  /**
   * The exponent for RSA keys.
   *
   * @example AQAB
   */
  e?: string | null;
  alg?: SignatureAlgorithm;
  /**
   * The intended use of the key (e.g., sig, enc).
   *
   * @example sig
   */
  use?: string | null;
  /**
   * A URL that points to an X.509 public key certificate or certificate chain.
   *
   * @format uri
   * @example https://example.com/cert.pem
   */
  x5u?: string | null;
  /**
   * The X.509 certificate chain.
   */
  x5c?: string[] | null;
  /**
   * The SHA-1 thumbprint of the X.509 certificate.
   *
   * @example dGhpcyBpcyBqdXN0IGEgdGh1bWJwcmludA
   */
  x5t?: string | null;
  /**
   * The SHA-256 thumbprint of the X.509 certificate.
   *
   * @example sM4KhEI1Y2Sb6-EVr6tJabmJuoP-ZE...
   */
  x5tS256?: string | null;
};
// ...
export type Jwks = {
  jwks: Jwk[];
};
// ...
export type AddSubordinate = {
  entity_id: EntityID;
  /**
   * Entity types the subordinate is registered for.
   */
  registered_entity_types?: string[];
  jwks?: Jwks;
};